0.20 Phase 4.7: split wasip2_tcp into directory + sub-bundles + wireup extract
Three-in-one internal refactor on the Tcp.* helper family. Zero
behaviour change — same wasm output, same emit order, same gating.

- src/codegen/wasm_gc/wasip2_tcp.rs (1945 lines) → directory with
  mod.rs / connect.rs / ids.rs / io.rs / lifecycle.rs. restore_bump
  lives once in mod.rs and every submodule imports it via
  `use super::restore_bump`.
- TcpConnectHelperFns (27 flat fields) → four sub-bundles
  TcpConnectPool / TcpConnectDns / TcpConnectSocket /
  TcpConnectMaterialize, grouped by pipeline concern. Callsites
  read `helpers.pool.tcp_pool_global`, `helpers.dns.poll_fn`, etc.
- Tcp.* type-section / fn-idx allocation pulled out of module.rs
  into wasip2_tcp/wireup.rs. emit_module_with now drops two
  one-liners (`wireup::allocate(...)` + `wireup::register_funcs(...)`)
  in place of the 420 inline lines, leaving body emission inline
  because it pulls in module-private structs.

module.rs shrinks by 388 lines net (130 added / 518 removed).
623 lib tests + 5 wasip2_tcp + 7 wasip2_tcp_stress + 6 wasip2_stress
green; clippy -D warnings + cargo fmt clean.
